Add files via upload
Contains two methods, a circulate method and an alter method. Circulate will create the arraylist from a string and circulate the letters. The alter method will add to each character in the arraylist. Circulate returns an arraylist of characters and accepts a string, integer, and a letter to indicate direction as parameters. The alter method accepts an arraylist as a parameter and alters that arraylist. It is a void method. Circulate calls the alter method.