Add files via upload
This program reads in five strings from a file. The strings contain a word and a key and then creates a cipher for each string. The cipher is made by converting the word into binary using ascii code, converting the key into binary, XOR the key and each letter (into ascii code), and printing the resulting hex code.